# functions/src/api/oas-doc.yaml
openapi: 3.0.0
info:
  title: Gom Don API
  description: Gom Don API.
  version: "1.0"
  contact:
    name: "Huynh Nhọn"
    email: "huynhnhon.dev@gmail.com"
paths:
  /buy:
    post:
      x-swagger-router-controller: buyRoute
      operationId: createBuy
      tags:
        - /buy
      description: CTVmua tạo một đơn mua trên 1688. Thông tin chung của đơn mua được lưu tại bảng buys. Thông tin về các phân loại được mua được lưu tại bảng buyModels
      summary: Tạo đơn nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyCreate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
    put:
      x-swagger-router-controller: buyRoute
      operationId: updateBuy
      tags:
        - /buy
      description: Cập nhật thông tin chung của đơn mua (Thông tin trong bảng buys)
      summary: Cập nhật đơn mua
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /buy/get-by-ids:
    get:
      x-swagger-router-controller: buyRoute
      operationId: getByIdsBuy
      description: Xem đơn mua theo danh sách id
      summary: get buy by array id
      tags:
        - /buy
      parameters:
        - in: query
          name: ids
          description: Là mảng bắt buộc. Mỗi phần tử trong mảng là id của đơn mua được lấy trên 1688
          required: true
          schema:
            uniqueItems: true
            type: string
            items:
              type: string
              maxItems: 40
              minItems: 1
          explode: false
          style: form
        - in: query
          name: fields
          description: Không bắt buộc. Mỗi phần tử trong mảng là tên của thuộc tính muốn xem. Sẽ không trả về thông tin nếu mảng trống. Nếu không gửi mảng lên server sẽ lấy về thông tin đầy đủ
          schema:
            uniqueItems: true
            type: string
            items:
              type: string
          explode: false
          style: form
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponsesWithArrayData"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
...
  /buy/get-by-status:
    get:
      x-swagger-router-controller: buyRoute
      operationId: getByStatusBuy
      description: Xem danh sách đơn mua theo trạng thái và id của quanlyCTVmua
      summary: Xem danh sách đơn mua theo trạng thái
      tags:
        - /buy
      parameters:
        - in: query
          name: status
          schema:
            uniqueItems: true
            type: string
            items:
              type: number
          allowEmptyValue: false
          required: true
          style: form
          explode: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /buy-model/gomdon-buy-id/{gomdon_buy_id}:
    get:
      x-swagger-router-controller: buyModelRouter
      operationId: getModelOfBuy
      tags:
        - /buy-model
      description: Xem toàn bộ buyModel của một đơn mua
      parameters:
        - in: path
          required: true
          name: gomdon_buy_id
          description: gomdon_buy_id là id document của đơn mua
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /buy-model:
    put:
      x-swagger-router-controller: buyModelRouter
      operationId: updateBuyModel
      tags:
        - /buy-model
      description: Cập nhật thông tin phân loại của đơn mua
      summary: Cập nhật thông tin phân loại của đơn mua
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyModelUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /product:
    post:
      x-swagger-router-controller: productRoute
      operationId: createProduct
      tags:
        - /product
      description: Thêm một sản phẩm mới. Cùng một sản phẩm có thể được thêm từ 2 nhân viên khác nhau. Tức một sản phẩm được định nghĩa dựa trên id sản phẩm và id nhân viên (gomdon_id, gomdon_by)
      summary: Thêm sản phẩm mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
    put:
      x-swagger-router-controller: productRoute
      operationId: updateProduct
      tags:
        - /product
      description: API này dùng để cập nhật sản phẩm. Lưu ý classify của sản phẩm được sẽ thay mới hoàn toàn nếu Dữ liệu gửi lên có classify
      summary: Cập nhật sản phầm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /product/delete/{product_id}:
    get:
      x-swagger-router-controller: productRoute
      operationId: deleteProduct
      tags:
        - /product
      description: API này dùng để xóa sản phẩm
      parameters:
        - in: path
          required: true
          name: product_id
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /product/get-all:
    get:
      x-swagger-router-controller: productRoute
      operationId: getAllProduct
      tags:
        - /product
      description: Xem tất cả sản phẩm
      summary: Xem tất cả sản phẩmm
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /product/get-by-ids:
    get:
      x-swagger-router-controller: productRoute
      operationId: getByIds
      tags:
        - /product
      description: người dùng gửi lên danh sách id của sản phẩm trên trang 1688. server trả về thông tin chi tiết các sản phẩm mà người dùng đó tạo theo id được gửi.
      summary: Xem chi tiết một sản phẩm theo id
      parameters:
        - in: query
          name: linked_product_ids
          description: id sản phẩm trên trang 1688
          required: true
          schema:
            type: string
            items:
              type: string
              maxItems: 40
              minItems: 1
          explode: false
          style: form
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponsesWithArrayData"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
...
  /export/create:
    post:
      x-swagger-router-controller: exportRoute
      operationId: createExport
      tags:
        - /export
      description: Phiếu xuất
      summary: Phiếu xuất
      requestBody:
        required: true
        content:
          application/json:
            schema:
              additionalProperties: false
              type: array
              uniqueItems: true
              items:
                type: object
                properties:
                  order_sn:
                    type: string
                    description: Mã đơn
                  shipping_traceno:
                    type: string
                    description: Mã vận đơn
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /export/get-by-status:
    get:
      x-swagger-router-controller: exportRoute
      operationId: :getByStatus
      description: get export by status
      summary: get export by status
      tags:
        - /export
      parameters:
        - in: query
          name: status
          schema:
            uniqueItems: true
            type: string
            items:
              type: number
          allowEmptyValue: false
          required: true
          explode: true
          style: form
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /export/{export_id}:
    get:
      x-swagger-router-controller: exportRoute
      operationId: getByIdExport
      tags:
        - /export
      parameters:
        - in: path
          required: true
          name: export_id
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
        "500":
          description: error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responses"
  /export/update:
    put:
      x-swagger-router-controller: exportRoute
      operationId: updateExport
      description: Cập nhật trạng thái phiếu xuất
      summary: Cập nhật trạng thái phiếu xuất
      tags:
        - /export
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - data
                - export_id
              properties:
                export_id:
                  type: string
                data:
                  type: object
                  minProperties: 1
                  properties:
                    gomdon_status:
                      type: number
                      enum:
                        - 1
                        - 2
                        - 3
                        - 4
      responses:
        "200":
          description: OK
  /export/add-snapshot:
    post:
      x-swagger-router-controller: exportRoute
      operationId: addSnapshotExport
      description: Thêm snapshot export
      summary: Thêm snapshot export
      tags:
        - /export
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, export_id]
              properties:
                image:
                  type: string
                  format: binary
                export_id:
                  type: string
      responses:
        "200":
          description: OK
        "500":
          description: error
...
  /sell/get-by-status:
    get:
      x-swagger-router-controller: sellRoute
      operationId: getByStatus
      description: get sell by shop_id and status
      summary: get sell by shop_id and status
      tags:
        - /sell
      parameters:
        - in: query
          name: status
          schema:
            uniqueItems: true
            type: string
            items:
              type: number
          allowEmptyValue: false
          required: true
          style: form
          explode: false
        - in: query
          name: shop_id
          schema:
            type: number
          allowEmptyValue: false
          # required: true
        - in: query
          name: field
          schema:
            type: string
            items:
              type: string
          allowEmptyValue: true
          style: form
          explode: false
      responses:
        "200":
          description: OK
          # content:
          #   application/json:
          #     schema:
          #       $ref: "#/components/schemas/Responses"
...
servers:
  - url: /
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GomdonLogs:
      type: object
      required: ["content", "id", "name", "type", "uid"]
      properties:
        content:
          type: string
        id:
          type: number
        name:
          type: string
        type:
          type: number
        uid:
          type: string
...
    Product:
      description: Dữ liệu chuẩn của bảng Product
      type: object
      additionalProperties: false
      # required:
      #   [
      #     classify,
      #     gomdon_product_name,
      #     linked_product_id,
      #     linked_company_name,
      #     images_preview,
      #   ]
      properties:
        gomdon_by:
          type: string
        classify:
          description: mảng các phân loại của sản phẩm
          type: array
          uniqueItems: true
          minItems: 1
          items:
            $ref: "#/components/schemas/ProductClassify"
        gomdon_ctime:
          type: number
        gomdon_product_name:
          type: string
          description: Tên tiếng việt của sản phẩm
        images_preview:
          type: string
        linked_company_name:
          type: string
          description: Tên nhà bán trên 1688
        linked_product_id:
          type: string
          description: Id của sản phẩm trên trang 1688
        gomdon_id:
          type: string
        gomdon_sku:
          type: string
        name_china:
          type: string
          description: Tên trung quốc của sản phẩm (lấy trên 1688)
    ProductUpdate:
      description: Model này dùng để validate product khi gửi lên để update
      required: [gomdon_id, data]
      additionalProperties: false
      properties:
        gomdon_id:
          type: string
        data:
          type: object
          additionalProperties: false
          minItems: 1
          properties:
            classify:
              type: array
              uniqueItems: true
              minItems: 1
              items:
                $ref: "#/components/schemas/ProductClassify"
            gomdon_product_name:
              type: string
            image_preview:
              type: string
    ProductClassify:
      additionalProperties: false
      type: object
      required: [gomdon_name, image]
      properties:
        gomdon_name:
          type: string
          description: Tên tiếng việt của phân loại
        gomdon_sku:
          type: string
          description: sku này do mình quản lý
        sku_name:
          type: string
          description: tên trung quốc của phân loại lấy trên 1688
        skuId:
          type: string
          description: skuid của phân loại trên 1688
        image:
          type: string
        length:
          type: number
        width:
          type: number
        height:
          type: number
        weight:
          type: number
        specId:
          type: string
        saleCount:
          type: number
        discountPrice:
          type: number
        canBookCount:
          type: number
    Buy:
      type: object
      additionalProperties: false
      required: [shipping_fee, company_name]
      properties:
        gomdon_by:
          type: object
          properties:
            CTVmua:
              type: string
            quanlyCTVmua:
              type: string
        gomdon_logs:
          type: array
          items:
            $ref: "#/components/schemas/GomdonLogs"
        gomdon_product_ids:
          type: array
          items:
            type: string
          description: Mảng các id sản phẩm được lấy trên trang 1688. Mảng này được lấy ra từ BuyModel để biết đơn mua này có mấy sản phẩm và nó là sản phầm gì
        origin_price:
          type: number
          description: Giá phải trả cho nhà bán (Không tính giá phát sinh như phí vận chuyển ...) Được tính từ bảng buyModel
        product:
          type: object
          additionalProperties: false
          properties:
            gomdon_product_name:
              type: string
            images_preview:
              type: string
        models:
          type: array
          items:
            type: object
          description: Phân loại hàng trong đơn
        currency:
          type: string
          description: Đơn vị tiền tệ (lấy trên 1688)
        currency_rate:
          type: number
          description: Tỉ giá
        note:
          type: string
        shipping_fee:
          type: number
          description: Phí ship từ nhà bán bên trung quốc về kho trung gian bên trung quốc(được lấy trên đơn mua 1688)
        voucher_price:
          type: number
          description: Số tiền được giảm (được lấy trên 1688)
        shipping_traceId:
          type: array
          items:
            type: string
        status:
          type: number
        company_name:
          type: string
          description: Tên nhà bán trên 1688
        gomdon_ctime:
          type: number
        gomdon_id:
          type: string
          description: Id document do mình quản lý
        buy_id:
          type: string
          description: Id của đơn mua trên 1688
    BuyModelUpdate:
      type: object
      additionalProperties: false
      required: [gomdon_product_id, gomdon_sku]
      properties:
        gomdon_product_id:
          type: string
          description: id san pham
        gomdon_sku:
          type: string
        gomdon_name:
          type: string
          description: ten phan loai
        gomdon_buy_id:
          type: string
          description: id don nhap
        actual_quantity:
          type: number
          description: thuc nhap
    BuyModel:
      type: object
      additionalProperties: false
      required:
        [
          linked_product_id,
          gomdon_sku,
          price,
          quantity,
          gomdon_product_id,
          skuId,
        ]
      properties:
        gomdon_id:
          type: string
          description: Id của document do mình quản lý
        gomdon_buy_id:
          type: string
          description: Id của đơn mua (Được liên kết đến trường gomdon_id của bảng Buy)
        gomdon_status:
          type: number
          enum:
            - 1
            - 2
            - 3
        actual_quantity:
          type: number
          description: Số lượng thực tế nhận được
        linked_product_id:
          type: string
          description: Id của sản phẩm trên trang 1688 (Có thể lấy từ bảng sản phẩm)
        image:
          type: string
          description: Lấy từ bảng sản phẩm
        gomdon_name:
          type: string
          description: Tên tiếng việt của phân loại
        gomdon_sku:
          type: string
          description: link đến gomdon_sku của ProductClassify
        price:
          type: number
          description: giá nhập
        quantity:
          type: number
          description: Số lượng của phân loại (lấy trên đơn mua của 1688)
        height:
          type: number
        length:
          type: number
        gomdon_product_id:
          type: string
          description: link đến gomdon_id của bảng products
        gomdon_product_name:
          type: string
          description: Tên tiếng việt của sản phẩm
        width:
          type: number
        weight:
          type: number
        sku_name:
          type: string
          description: Tên phân loại trên 1688
        skuId:
          type: string
          description: sku phân loại trên 1688
        stock:
          type: number
          description: Số lượng hàng có trong kho
    BuyCreate:
      additionalProperties: false
      required: [buy, buyModel]
      properties:
        buy:
          $ref: "#/components/schemas/Buy"
          description: Thông tin chung của đơn nhập. Thông tin này được lưu tại bảng Buys
        buyModel:
          type: array
          description: Thông tin về các phân loại hàng được mua trong đơn nhập. Được lưu tại bảng buy_model. Dữ liệu này được sử dụng để làm bảng kho
          items:
            $ref: "#/components/schemas/BuyModel"
    BuyUpdate:
      additionalProperties: false
      required: [gomdon_id, data]
      properties:
        buy_model_ids:
          type: array
          items:
            type: object
            properties:
              gomdon_id:
                type: string
                description: id phan loai hang
              actual_quantity:
                type: number
                description: so luong thuc nhap cua phan loai
        gomdon_id:
          type: string
          description: id document của bảng buys
        data:
          type: object
          minProperties: 1
          properties:
            currency_rate:
              type: number
              description: Tỉ giá
            note:
              type: string
            shipping_traceId:
              type: array
              items:
                type: string
            status:
              type: number
...
    ResponsesWithArrayData:
      type: array
      items:
        type: object
        required: [id, code, data]
        properties:
          id:
            type: string
          code:
            type: string
            enum:
              - success
              - error
          data:
            type: object
    Responses:
      type: object
      required:
        - code
        - data
      properties:
        code:
          type: string
          enum:
            - success
            - error
        data:
          oneOf:
            - type: array
              items:
                type: object
            - type: object
            - type: string
security:
  - Bearer: []